requirements.txt
tensorflow
Pillow
matplotlib
flask